+++ App.js
@@ -35,6 +35,13 @@ export default function App() {
     "You spent 3 hours at Brittan Dining Hall",
     "You spent 3 hours at Brittan Dining Hall",
     "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
+    "You spent 3 hours at Brittan Dining Hall",
     "You spent 4 hours at the Georgia Tech Library",
     "You spent 2 hours at Clough",
     "You spent 5 hours at Klaus",
